pyproject.toml
[project]
name = "ecom-support-agent"
version = "0.1.0"
description = "Customer Support Agent for E-Commerce using LangGraph"
requires-python = ">=3.10"

dependencies = [
  "langgraph",
  "langchain-core",
  "groq",
  "python-dotenv",
  "pydantic",
  "pytest",
  "supabase",
  "requests"
]

